// src/utils/enum.ts
export enum COLORS {
  BLACK = "#000000",
  WHITE = "#FFFFFF",
  MINT_GREEN = "#9e2b25",
  BANNER_COLOR = "#EAF0F2",
  PRIMARY = "#CE0809",
  TEXT = "#727695",
  SECONDARY = "#3EBF78",
  BLUE = "#0E2742",
  GREY = "#EFF3FF",
  SECTIONBG = "#E7F0FD",
  FOOTERBG = "#12141B",
  TRANSPARENT = "transparent",
  LIGHT_BLUE = "#BDD7E7",
  FAQBG = "#F2F2FE",
}

export enum HEADER_LINKS {
  HOME = "Home",
  ABOUT = "About",
  PROGRAMS = "Programs",
  CONTACT = "Contact",
  COMBOPACKS = "Combo Packs",
}

export enum COURSE_TYPE {
  TECH = "TECH & DATA",
  BUSINESS = "BUSINESS",
}

export enum COURSE_IDS {
  WEB_DEVELOPMENT = "web-development",
  ARTIFICIAL_INTELLIGENCE = "artificial-intelligence",
  DATA_SCIENCE = "data-science",
  MACHINE_LEARNING = "machine-learning",
  MOBILE_APP_DEVELOPMENT = "mobile-app-development",
  DIGITAL_MARKETING = "digital-marketing",
  FINANCE = "finance",
  HUMAN_RESOURCE = "human-resource",
}